refactore app entry point to follow best practices
